# api/openapi.yaml
openapi: 3.0.0
info:
  contact: {}
  description: Configure and inspect a running TDNS server.
  title: TDNS Management API
  version: "1.0"
paths:
...
  /api/cache:
    delete:
      description: Delete all currently cached DNS responses.
      operationId: cacheClear
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                type: string
      security:
        - BearerAuth: []
        - CookieAuth: []
      summary: Clear the DNS cache
      tags:
        - cache
    get:
      description: Return cache configuration and counters.
      operationId: cacheStatus
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                type: string
      security:
        - BearerAuth: []
        - CookieAuth: []
      summary: Get cache status
      tags:
        - cache
  "/api/cache/{action}":
    post:
      description: Enable or disable the cache and persist the selected state.
      operationId: cacheToggle
      parameters:
        - description: Requested state
          in: path
          name: action
          required: true
          schema:
            type: string
            enum:
              - start
              - stop
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                type: string
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
      security:
        - BearerAuth: []
        - CookieAuth: []
      summary: Toggle the DNS cache
      tags:
        - cache
  /api/cache/excludes:
    post:
      description: Replace and persist selectors excluded from caching.
      operationId: cacheExcludesReplace
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/api.CacheExcludeRequest"
        description: Cache exclusion selectors
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                type: string
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
      security:
        - BearerAuth: []
        - CookieAuth: []
      summary: Replace cache exclusions
      tags:
        - cache
  /api/dns-log/alias:
    post:
      description: Associate a client IP address with a display name.
      operationId: dnsLogAliasSet
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/api.DNSLogAliasRequest"
        description: Client alias
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                type: string
      security:
        - BearerAuth: []
        - CookieAuth: []
      summary: Set a DNS client alias
      tags:
        - dns-log
  /api/dns-log/clients:
    get:
      description: Search observed clients by address or alias.
      operationId: dnsLogClientsSearch
      parameters:
        - description: Address or alias substring
          in: query
          name: search
          schema:
            type: string
        - description: Maximum result count
          in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                type: string
      security:
        - BearerAuth: []
        - CookieAuth: []
      summary: Search DNS clients
      tags:
        - dns-log
  /api/dns-log/dashboard:
    get:
      description: Return summary and hourly query statistics.
      operationId: dnsLogDashboard
      parameters:
        - description: Dashboard window in hours
          in: query
          name: hours
          schema:
            type: integer
            minimum: 1
            maximum: 336
            default: 24
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                type: string
      security:
        - BearerAuth: []
        - CookieAuth: []
      summary: Get DNS log dashboard
      tags:
        - dns-log
  /api/dns-log/rotate:
    post:
      description: Delete DNS log entries selected by a relative duration.
      operationId: dnsLogRotate
      parameters:
        - description: Relative age such as 24h or 1w
          in: query
          name: since
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                type: string
      security:
        - BearerAuth: []
        - CookieAuth: []
      summary: Rotate DNS log entries
      tags:
        - dns-log
  "/api/dns-log/top/{top}":
    get:
      description: Return top domains with optional status and client filters.
      operationId: dnsLogTop
      parameters:
        - description: Maximum result count
          in: path
          name: top
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 50
        - description: Relative age such as 24h or 1w
          in: query
          name: since
          schema:
            type: string
        - description: Query disposition
          in: query
          name: status
          schema:
            type: string
            enum:
              - blocked
              - allowed
        - description: Client alias or IP address
          in: query
          name: client
          schema:
            type: string
        - description: Client matching mode
          in: query
          name: client_mode
          schema:
            type: string
            enum:
              - host
              - ip
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                type: string
      security:
        - BearerAuth: []
        - CookieAuth: []
      summary: Get top queried domains
      tags:
        - dns-log
...
servers:
  - url: /
components:
...
  securitySchemes:
    BearerAuth:
      description: Bearer token using the format "Bearer {token}".
      in: header
      name: Authorization
      type: apiKey
    CookieAuth:
      description: Opaque HttpOnly browser session cookie set by POST /api/auth/exchange.
      in: cookie
      name: __Host-tdns-session
      type: apiKey
  schemas:
...
    api.BlacklistStatus:
      properties:
        blockfile_total_entries:
          type: integer
        enabled:
          type: boolean
        excludes:
          items:
            type: string
          type: array
        external_file:
          type: string
        external_pull_period:
          type: string
        external_repo:
          type: string
        external_repo_branch:
          type: string
        file:
          type: string
        persisted_excludes:
          items:
            type: string
          type: array
        persisted_hosts:
          items:
            type: string
          type: array
        runtime_whitelist:
          items:
            type: string
          type: array
      type: object
...
    api.CacheExcludeRequest:
      properties:
        excludes:
          items:
            type: string
          type: array
      type: object
    api.CacheStatus:
      properties:
        enabled:
          type: boolean
        excludes:
          items:
            type: string
          type: array
        hits:
          type: integer
        misses:
          type: integer
        ttl:
          type: integer
      type: object
    api.ClientCandidate:
      properties:
        address:
          example: 192.0.2.10
          type: string
        host:
          example: office
          type: string
      type: object
    api.DNSLogAliasRequest:
      properties:
        addr:
          example: 192.0.2.10
          type: string
        name:
          example: office
          type: string
      type: object
    api.DashboardHourlyPoint:
      properties:
        allowed_queries:
          type: integer
        blocked_queries:
          type: integer
        hour_bucket:
          type: integer
        hour_start:
          type: string
        total_queries:
          type: integer
      type: object
    api.DashboardSummary:
      properties:
        allowed_queries:
          type: integer
        blocked_queries:
          type: integer
        cache_hits:
          type: integer
        cache_misses:
          type: integer
        total_queries:
          type: integer
      type: object
...
    api.HostEntry:
      properties:
        address:
          example: 192.0.2.10
          type: string
        domain:
          example: example.com
          type: string
      type: object
    api.KnownHost:
      properties:
        address:
          example: 192.0.2.10
          type: string
        host:
          example: office
          type: string
      type: object
    api.LogDetails:
      properties:
        counter:
          example: 42
          minimum: 0
          type: integer
        domain:
          example: example.com.
          type: string
        host:
          example: office
          type: string
      type: object
...
    api.Response:
      properties:
        blacklist:
          $ref: "#/components/schemas/api.BlacklistStatus"
        cache:
          $ref: "#/components/schemas/api.CacheStatus"
        clients:
          items:
            $ref: "#/components/schemas/api.ClientCandidate"
          type: array
        current_status:
          example: enabled
          type: string
        hourly:
          items:
            $ref: "#/components/schemas/api.DashboardHourlyPoint"
          type: array
        items:
          items:
            type: string
          type: array
        kind:
          example: api.tdns/cache/response
          type: string
        known_hosts:
          items:
            $ref: "#/components/schemas/api.KnownHost"
          type: array
        log_items:
          items:
            $ref: "#/components/schemas/api.LogDetails"
          type: array
        message:
          example: Status OK
          type: string
        static_response:
          $ref: "#/components/schemas/api.StaticResponseStatus"
        stub_resolver:
          $ref: "#/components/schemas/api.StubResolverStatus"
        summary:
          $ref: "#/components/schemas/api.DashboardSummary"
        tag_members:
          items:
            $ref: "#/components/schemas/api.TagMember"
          type: array
        window_hours:
          maximum: 336
          minimum: 1
          type: integer
        zen_mode:
          $ref: "#/components/schemas/api.ZenModeStatus"
      type: object
...
    api.StaticResponseStatus:
      properties:
        configured_hosts:
          items:
            $ref: "#/components/schemas/api.HostEntry"
          type: array
        enabled:
          type: boolean
        file:
          type: string
        labels:
          items:
            type: string
          type: array
        persisted_hosts:
          items:
            $ref: "#/components/schemas/api.HostEntry"
          type: array
        runtime_hosts:
          items:
            $ref: "#/components/schemas/api.HostEntry"
          type: array
      type: object
...
    api.StubResolverStatus:
      properties:
        configured_stubs:
          items:
            type: string
          type: array
        enabled:
          type: boolean
        runtime_stubs:
          items:
            type: string
          type: array
      type: object
    api.TagMember:
      properties:
        address:
          example: 192.0.2.10
          type: string
        has_host_alias:
          type: boolean
        host:
          example: office
          type: string
      type: object
...
    api.ZenModeStatus:
      properties:
        configured_domains:
          items:
            type: string
          type: array
        configured_excludes:
          items:
            type: string
          type: array
        duration_minutes:
          type: integer
        enabled:
          type: boolean
        ends_at:
          type: string
        file:
          type: string
        labels:
          items:
            type: string
          type: array
        persisted_domains:
          items:
            type: string
          type: array
        persisted_excludes:
          items:
            type: string
          type: array
        remaining_seconds:
          type: integer
        runtime_domains:
          items:
            type: string
          type: array
        started_at:
          type: string
      type: object
...